+++ src/components/ChartBuilder/index.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import ChartCanvas from './ChartCanvas.vue'
+import InteractionCanvas from './InteractionCanvas.vue'
 import TopBar from './TopBar.vue'
 </script>
 
@@ -7,7 +8,10 @@ import TopBar from './TopBar.vue'
   <div class="chart-builder">
     <div class="chart-builder-contents">
       <TopBar />
-      <ChartCanvas />
+      <div class="canvas-container">
+        <ChartCanvas id="chart-canvas" />
+        <InteractionCanvas id="interaction-canvas" />
+      </div>
     </div>
   </div>
 </template>
@@ -23,6 +27,23 @@ import TopBar from './TopBar.vue'
 .chart-builder-contents {
   margin: 0 auto;
   padding: 0;
+}
+
+.canvas-container {
+  display: grid;
+}
+
+.canvas-container canvas {
+  grid-column: 1;
+  grid-row: 1;
+}
+
+#chart-canvas {
+  z-index: 1;
+}
+
+#interaction-canvas {
+  z-index: 2;
 }
 
 @media screen and (max-width: 1000px) {
